linux/sigsetops: fix type confusion (bug 31468)
Each mask in the sigset array is an unsigned long, so fix __sigisemptyset
to use that instead of int.  The __sigword function returns a simple array
index, so it can return int instead of unsigned long.

+++ sysdeps/unix/sysv/linux/internal-sigset.h
@@ -54,7 +54,7 @@ static inline int
 internal_sigisemptyset (const internal_sigset_t *set)
 {
   int cnt = __NSIG_WORDS;
-  int ret = set->__val[--cnt];
+  unsigned long int ret = set->__val[--cnt];
   while (ret == 0 && --cnt >= 0)
     ret = set->__val[cnt];
   return ret == 0;
@@ -82,7 +82,7 @@ static inline int
 internal_sigismember (const internal_sigset_t *set, int sig)
 {
   unsigned long int mask = __sigmask (sig);
-  unsigned long int word = __sigword (sig);
+  int word = __sigword (sig);
   return set->__val[word] & mask ? 1 : 0;
 }
 
@@ -90,7 +90,7 @@ static inline void
 internal_sigaddset (internal_sigset_t *set, int sig)
 {
   unsigned long int mask = __sigmask (sig);
-  unsigned long int word = __sigword (sig);
+  int word = __sigword (sig);
   set->__val[word] |= mask;
 }
 
@@ -98,7 +98,7 @@ static inline void
 internal_sigdelset (internal_sigset_t *set, int sig)
 {
   unsigned long int mask = __sigmask (sig);
-  unsigned long int word = __sigword (sig);
+  int word = __sigword (sig);
   set->__val[word] &= ~mask;
 }
